Use event checking instead of checking user agent
This method checks that the “ontouchforcechange” event exists inside
the document. This means you don’t have to use the user agent checking
because that event is only available in iOS 10 or higher.

This is also more explicit because you only want to run the older
timeout method if the new “touchforcechange” is not called.